requirements.txt
absl-py==2.3.1
addict==2.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
camtools==0.1.7
certifi==2022.12.7
#chamfer_3D==0.0.0
charset-normalizer==2.1.1
click==8.1.8
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==3.2.0
decorator==5.2.1
diffusers==0.35.2
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.13.1
Flask==3.1.1
fonttools==4.59.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.1
h5py==3.14.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.4
imageio==2.13.0
imageio-ffmpeg==0.4.8
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lightning-utilities==0.15.2
lpips==0.1.4
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
open3d==0.18.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
plotly==6.3.0
portalocker==3.2.0
prompt_toolkit==3.0.51
protobuf==6.32.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
-e git+https://github.com/facebookresearch/pytorch3d.git@fc6a6b895185a502c4064fc5ea7ecb760fd1a0d6#egg=pytorch3d
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.28.1
retrying==1.4.2
rich==14.1.0
rpds-py==0.27.0
safetensors==0.6.2
scikit-image==0.19.3
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.42.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.19
#tinycudann==2.0
torch==2.1.0+cu118
torch-ema==0.3
torchaudio==2.1.0+cu118
torchmetrics==1.8.1
torchvision==0.16.0+cu118
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
triton==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.13
wandb==0.22.2
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.23.0
